requirements.txt
streamlit
pandas
numpy
plotly
yfinance
python-dotenv
requests
lxml
finnhub-python
lightgbm
scikit-learn
joblib
nltk
vaderSentiment
deep-translator
pyyaml
optuna
websocket-client
